requirements.txt
aiogram==3.4.1
fastapi==0.110.0
uvicorn[standard]==0.27.1
openai>=1.0.0
gspread==5.12.0
oauth2client==4.1.3
python-dotenv==1.0.1